Se agrega caso de uso para obtener resumen ejecutivo

+++ src/main/java/com/aquavitae/application/usecase/RiesgoHidricoFacadePort.java
@@ -13,5 +13,5 @@ public interface RiesgoHidricoFacadePort {
     List<PlantaRiesgo> obtenerPlantasConRiesgo();
     long contarCrisisActivas();
     List<AlertaResumen> obtenerAlertasRecientes(int limite);
-    Map<String, Double> obtenerEvolucion7Dias();
+    Map<String, Float> obtenerEvolucion7Dias();
 }
